+++ src/containers/NewContractMember/NewContractMemberSummary.jsx
@@ -21,10 +21,10 @@ import TermsDialog from '../../components/TermsDialog'
 import LegalText from '../../components/LegalText'
 
 import { iconRequirements } from '../../themes/commonStyles'
-import { NEW_MEMBER_CONTRACT_FORM_SUBSTEPS } from '../../services/steps'
 import { getPrices } from '../../services/api'
 import { THOUSANDS_CONVERSION_FACTOR } from '../../services/utils'
 
+import { NEW_MEMBER_CONTRACT_FORM_SUBSTEPS, NEW_LINK_MEMBER_CONTRACT_FORM_SUBSTEPS } from '../../services/steps'
 import Loading from '../../components/Loading'
 import PlaceOutlinedIcon from '@mui/icons-material/PlaceOutlined'
 import SettingsOutlinedIcon from '@mui/icons-material/SettingsOutlined'
@@ -53,6 +53,12 @@ const NewContractMemberSummary = (props) => {
   const [prices, setPrices] = useState({})
   const [openGeneralTermsDialog, setOpenGeneralTermsDialog] = useState(false)
   const [showReviewLinks, setShowReviewLinks] = useState(false)
+
+  const formSteps = values?.has_member == 'member-off'
+    ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS
+    : values?.has_member == 'member-on'
+    ? NEW_LINK_MEMBER_CONTRACT_FORM_SUBSTEPS
+    : undefined
 
   const handleChangePrivacyPolicy = (event) => {
     const checked = event.target.checked
@@ -83,100 +89,162 @@ const NewContractMemberSummary = (props) => {
     gl_ES: 'Galego'
   }
 
-  const legalReviewFields = {
-    icon: <PersonIcon sx={iconRequirements} />,
-    title: t('REVIEW_HOLDER_TITLE'),
-    field: [
+  const legalReviewFields = [
       {
         reviewLabel: t('BUSINESS_NAME'),
         reviewValue: values?.new_member?.name,
         step: showReviewLinks
-          ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['MEMBER_INFO']
+          ? formSteps['MEMBER_INFO']
           : null
       },
       {
         reviewLabel: t('CIF'),
         reviewValue: values?.new_member?.nif,
         step: showReviewLinks
-          ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['IDENTIFY_MEMBER']
+          ? values?.has_member == 'member-off'
+            ? formSteps['IDENTIFY_MEMBER']
+            : formSteps['MEMBER_INFO']
           : null
       },
       {
         reviewLabel: t('PROXY'),
         reviewValue: `${values?.new_member?.proxyname} (${values?.new_member?.proxynif})`,
         step: showReviewLinks
-          ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['MEMBER_INFO']
+          ? formSteps['MEMBER_INFO']
           : null
       },
       {
         reviewLabel: t('REVIEW_SUPPLY_POINT_LABEL_ADDRESS'),
         reviewValue: `${values?.address?.street} ${values?.address?.number}`,
         step: showReviewLinks
-          ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['MEMBER_INFO']
+          ? formSteps['MEMBER_INFO']
           : null
       },
       {
         reviewLabel: t('REVIEW_SUPPLY_POINT_LABEL_CITY'),
         reviewValue: values?.address?.city?.name,
         step: showReviewLinks
-          ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['MEMBER_INFO']
+          ? formSteps['MEMBER_INFO']
           : null
       }
     ]
-  }
 
-  const physicalReviewFields = {
-    icon: <PersonIcon sx={iconRequirements} />,
-    title: t('REVIEW_HOLDER_TITLE'),
-    field: [
+  const physicalReviewFields = [
       {
         reviewLabel: t('REVIEW_HOLDER_LABEL_NAME'),
         reviewValue: `${values?.new_member?.name} ${values?.new_member?.surname1} ${values?.new_member?.surname2}`,
         step: showReviewLinks
-          ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['MEMBER_INFO']
+          ? formSteps['MEMBER_INFO']
           : null
       },
       {
         reviewLabel: t('REVIEW_HOLDER_LABEL_NIF'),
         reviewValue: values?.new_member?.nif,
         step: showReviewLinks
-          ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['IDENTIFY_MEMBER']
+          ? values?.has_member == 'member-off'
+            ? formSteps['IDENTIFY_MEMBER']
+            : formSteps['MEMBER_INFO']
           : null
       },
       {
         reviewLabel: t('REVIEW_HOLDER_LABEL_PHONE'),
         reviewValue: `(${values?.new_member?.phone_code}) ${values?.new_member?.phone}`,
         step: showReviewLinks
-          ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['MEMBER_INFO']
+          ? formSteps['MEMBER_INFO']
           : null
       },
       {
         reviewLabel: t('REVIEW_HOLDER_LABEL_EMAIL'),
         reviewValue: values?.new_member?.email,
         step: showReviewLinks
-          ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['MEMBER_INFO']
+          ? formSteps['MEMBER_INFO']
           : null
       }
     ]
-  }
+
+  const oldHolderFields = [
+      {
+        reviewLabel: t('REVIEW_HOLDER_LABEL_NIF'),
+        reviewValue: values?.member?.nif,
+        step: showReviewLinks
+          ? formSteps['LINK_MEMBER']
+          : null
+      }
+    ]
+
+  const reviewHolderData = values?.member_is_holder == 'holder-member-yes'
+    ? oldHolderFields
+    : values?.new_member?.person_type == 'legal-person'
+    ? legalReviewFields
+    : physicalReviewFields
+
+  const summary_process = values?.previous_holder == 'prevuis-holder-yes'
+    ? t('NEW_CONTRACT_OLD_HOLDER_SUMMARY_PROCESS')
+    : t('NEW_CONTRACT_NEW_HOLDER_SUMMARY_PROCESS')
+
+  const reviewProcessFields = values?.has_member == 'member-off' ?
+    [
+      {
+        reviewValue: t('NEW_MEMBER_SUMMARY_PROCESS')
+      },
+      {
+        reviewValue: summary_process
+      }
+    ] : [
+      {
+        reviewValue: summary_process
+      },
+      {
+        reviewLabel: t('REVIEW_LINK_PERSON'),
+        reviewValue: values?.member?.nif,
+        step: showReviewLinks
+          ? formSteps['LINK_MEMBER']
+          : null
+      }
+    ]
+
+  const reviewContactData = values?.member_is_holder == 'holder-member-yes'
+    ? [
+      {
+        reviewValue: t('REVIEW_CONTACT_OLD_HOLDER')
+      }
+    ]
+    : [
+    {
+      reviewLabel: t('REVIEW_HOLDER_LABEL_PHONE'),
+      reviewValue: `(${values?.new_member?.phone_code}) ${values?.new_member?.phone}`,
+      step: showReviewLinks
+        ? formSteps['MEMBER_INFO']
+        : null
+    },
+    {
+      reviewLabel: t('REVIEW_HOLDER_LABEL_EMAIL'),
+      reviewValue: values?.new_member?.email,
+      step: showReviewLinks
+        ? formSteps['MEMBER_INFO']
+        : null
+    },
+    {
+      reviewLabel: t('LANGUAGE'),
+      reviewValue: languages[values?.new_member?.language],
+      step: showReviewLinks
+        ? formSteps['MEMBER_INFO']
+        : null
+    }
+  ]
 
   const reviewFields = [
     [
       {
         icon: <DescriptionOutlinedIcon sx={iconRequirements} />,
         title: t('REVIEW_PROCESS_TITLE'),
-        field: [
-          {
-            reviewValue: t('NEW_MEMBER_SUMMARY_PROCESS')
-          },
-          {
-            reviewValue: t('NEW_CONTRACT_SUMMARY_PROCESS')
-          }
-        ]
+        field: reviewProcessFields
       },
-      values?.new_member?.person_type == 'legal-person'
-        ? legalReviewFields
-        : physicalReviewFields
+      {
+        icon: <PersonIcon sx={iconRequirements} />,
+        title: t('REVIEW_HOLDER_TITLE'),
+        field: reviewHolderData
+      },
     ],
     [
       {
@@ -187,28 +255,28 @@ const NewContractMemberSummary = (props) => {
             reviewLabel: t('CUPS_LABEL'),
             reviewValue: values?.cups,
             step: showReviewLinks
-              ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['SUPPLY_POINT']
+              ? formSteps['SUPPLY_POINT']
               : null
           },
           {
             reviewLabel: t('REVIEW_SUPPLY_POINT_LABEL_ADDRESS'),
-            reviewValue: `${values?.address?.street} ${values?.address?.number}`,
+            reviewValue: `${values?.supply_point_address?.street} ${values?.supply_point_address?.number}`,
             step: showReviewLinks
-              ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['SUPPLY_INFO']
+              ? formSteps['SUPPLY_INFO']
               : null
           },
           {
             reviewLabel: t('REVIEW_SUPPLY_POINT_LABEL_CITY'),
-            reviewValue: values?.address?.city?.name,
+            reviewValue: values?.supply_point_address?.city?.name,
             step: showReviewLinks
-              ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['SUPPLY_INFO']
+              ? formSteps['SUPPLY_INFO']
               : null
           },
           {
             reviewLabel: 'CNAE',
             reviewValue: values?.supply_point.cnae,
             step: showReviewLinks
-              ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['SUPPLY_INFO']
+              ? formSteps['SUPPLY_INFO']
               : null
           }
         ]
@@ -223,9 +291,13 @@ const NewContractMemberSummary = (props) => {
           },
           {
             reviewLabel: t('FARE'),
-            reviewValue: values?.contract.tariff_mode,
+            reviewValue: isTariffIndexed
+              ? t('FARE_INDEXED')
+              : values.contract.power_type === 'power-higher-15kw'
+              ? t('FARE_PERIODS').concat(" ",'3.0TD')
+              : t('FARE_PERIODS').concat(" ",'2.0TD'),
             step: showReviewLinks
-              ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['MEMBER_INFO']
+              ? formSteps['MEMBER_INFO']
               : null
           },
           {
@@ -242,29 +314,7 @@ const NewContractMemberSummary = (props) => {
       {
         icon: <LocalPhoneOutlinedIcon sx={iconRequirements} />,
         title: t('REVIEW_CONTACT_INFORMATION_TITLE'),
-        field: [
-          {
-            reviewLabel: t('REVIEW_HOLDER_LABEL_PHONE'),
-            reviewValue: `(${values?.new_member?.phone_code}) ${values?.new_member?.phone}`,
-            step: showReviewLinks
-              ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['MEMBER_INFO']
-              : null
-          },
-          {
-            reviewLabel: t('REVIEW_HOLDER_LABEL_EMAIL'),
-            reviewValue: values?.new_member?.email,
-            step: showReviewLinks
-              ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['MEMBER_INFO']
-              : null
-          },
-          {
-            reviewLabel: t('LANGUAGE'),
-            reviewValue: languages[values?.new_member?.language],
-            step: showReviewLinks
-              ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['MEMBER_INFO']
-              : null
-          }
-        ]
+        field: reviewContactData
       },
       {
         icon: <CreditCardOutlinedIcon sx={iconRequirements} />,
@@ -274,21 +324,21 @@ const NewContractMemberSummary = (props) => {
             reviewLabel: t('REVIEW_PAYMENT_DATA'),
             reviewValue: t('REVIEW_PAYMENT_DATA_QUANTITY'),
             step: showReviewLinks
-              ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['PAYMENT_INFO']
+              ? formSteps['PAYMENT_INFO']
               : null
           },
           {
             reviewLabel: t('REVIEW_PAYMENT_DATA_LABEL_IBAN'),
             reviewValue: values?.new_member?.iban,
             step: showReviewLinks
-              ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['PAYMENT_INFO']
+              ? formSteps['PAYMENT_INFO']
               : null
           },
           {
             reviewLabel: t('VOLUNTARY_CENT'),
             reviewValue: values?.voluntary_donation ? t('YES') : t('NO'),
             step: showReviewLinks
-              ? NEW_MEMBER_CONTRACT_FORM_SUBSTEPS['DONATION']
+              ? formSteps['DONATION']
               : null
           }
         ]